chore: Update puppeteer-extra-plugin-stealth npm dependency to version 2.11.2
This commit updates the puppeteer-extra-plugin-stealth npm dependency from version 2.11.1 to version 2.11.2. The new version includes improvements and bug fixes that enhance the stealth capabilities of the browser setup. This update ensures that the code is using the latest stable version of the dependency and takes advantage of the latest features and enhancements.