refactor: Update save_every_check value to 1 in init.js
This commit updates the `save_every_check` value in the `init.js` file from 100 to 1. The previous value was causing infrequent saving of data, which was not desirable. By setting it to 1, the data will be saved after every check, ensuring that no data is lost and providing more accurate and up-to-date information.