refactor: Export makeFileStorage from init.js
This commit modifies the import and export statements in the code to export the `makeFileStorage` function from the `init.js` file. Previously, the `init.js` file was exporting the `init` function as the default export, but now it exports both the `init` and `makeFileStorage` functions as named exports. This change allows other modules to import and use the `makeFileStorage` function directly, improving code modularity and reusability.